Added New Scene for Selecting Songs
This new scene is allows for endless scrolling for songs. I need to add the ability for the Songs to be grabbed from the Songs folder, but for now, it is staticly created. I'm also implementating songs to now be heard in the Main scene. However, the delay doesn't seem to be working. Might be an issue with the current Unity version